Put and delete article API

+++ django-back/articles/views.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from .models import Article, Comment
 from .serializers import ArticleListSerializer, ArticleSerializer, CommentSerializer
-from .permissions import IsAuthorOrReadonly
+from .permissions import IsAuthor
 
 
 # Create your views here.
@@ -29,14 +29,16 @@ def article_create(request):
         return Response(serializer.data)
 
 @api_view(['PUT', 'DELETE'])
-@permission_classes([IsAuthorOrReadonly])
 def article_update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'PUT':
-        serializer = ArticleSerializer(data=request.data, instance=article)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+    if request.user.id == article.user.id:
+        if request.method == 'PUT':
+            serializer = ArticleSerializer(data=request.data, instance=article)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
+        else:
+            article.delete()
+            return Response({'message':'Article has been deleted!'})
     else:
-        article.delete()
-        return Response({'message':'Article has been deleted!'})
+        return Response({'message':'author only'})
